+++ src/main/java/com/KoreaIT/project/BAP/service/PointService.java
@@ -18,11 +18,11 @@ public class PointService {
 		this.pointRepository = pointRepository;
 	}
 
-	public void doWrite(int memberId, int payment_id, int p_point) {
+	public void doWrite(long memberId, int payment_id, int p_point) {
 		pointRepository.doWrite(memberId, payment_id, p_point);
 	}
 
-	public List<Point> getPointsByMemberId(int memberId) {
+	public List<Point> getPointsByMemberId(long memberId) {
 		return pointRepository.getPointsByMemberId(memberId);
 	}
 }
